put all correction and estimation function in a directory
